Changed profile disk creation method

+++ MigrateCitrixUPMtoFSLogix/functions/Convert-CFXUser.ps1
@@ -28,7 +28,7 @@
 
         # Set the user account as owner on FSLogix VHD and folder.
         [Parameter(Mandatory = $false)]
-        [ValidateScript({Get-ADGroup -Identity $_ -ErrorAction Stop})]
+        [ValidateScript( { Get-ADGroup -Identity $_ -ErrorAction Stop })]
         [String] $ADGroupName,
 
         # FSLogix Folder Name Pattern. You can use %SID% or %username%. Default is: "%Username%_%SID%"
@@ -55,17 +55,30 @@
         [ValidateScript( { Test-Path -Path $_ -PathType Container })]
         [string] $TempFolderPath = $env:Temp,
 
+        # Robocopy multithread count, default is 100 to maximize copy performance
+        [Parameter(Mandatory = $false)]
+        [int] $RobocopyMT = 100,
+
         # Do not delete temp VHD created on temp folder
         [Parameter(Mandatory = $false)]
         [switch] $DoNoCleanup
     )
     begin {
         if ($Username.count -gt 1 `
-            -and ($CitrixUPMProfilePath -NotLike "*%username%*") `
+                -and ($CitrixUPMProfilePath -NotLike "*%username%*") `
                 -or ($RedirectedFoldersPath -NotLike "*%username%*")
-            ) {
+        ) {
             throw "When providing multiple usernames, you must use %username% in parameters CitrixUPMProfilePath and RedirectedFoldersPath .`r`nFor example: \\FileServer\CTXProfiles\%username%\Win2016\UPM_Profile"
         }
+        if ($ADGroupName) {
+            try {
+                Import-Module -Name ActiveDirectory -ErrorAction Stop
+            }
+            catch {
+                throw "Could not load Active Directory module which is required when -ADGroupName is used."
+            }
+        }
+
     }
 
     process {
@@ -75,10 +88,11 @@
                 #region: Reset Variables
                 $sid = $null
                 $citrixProfilePath = $null
-                $UserRedirectedFoldersPath =$null
+                $userRedirectedFoldersPath = $null
                 $diskItemLength = $null
                 $copyResult = $null
                 $errorMessage = $null
+                $tempDiskCreated = $null
                 #endregion: Reset Variables
                 #region: Confirm username is correct and retrieve SID
                 $sid = Get-CFXADUserSID -Username $user -ErrorAction Stop
@@ -106,31 +120,68 @@
                 if ($Username.count -gt 1 -and $CitrixUPMProfilePath -NotLike "*%username%*") {
                     throw "When providing multiple usernames, you must use %username% in parameter RedirectedFoldersPath.`r`nFor example: \\FileServer\RedirectedFolders\%username%"
                 }
-                $UserRedirectedFoldersPath = Get-CFXProfilePath -Username $user -ProfilePath $RedirectedFoldersPath -ErrorAction Stop
+                $userRedirectedFoldersPath = Get-CFXProfilePath -Username $user -ProfilePath $RedirectedFoldersPath -ErrorAction Stop
                 #If folder not found, fail
 
                 #endregion: Calculate variables
 
                 #region: Validate user is not member of AD Group
-                if($ADGroupName){
+                if ($ADGroupName) {
                     Write-PSFMessage -Level Verbose -Message "Confirming $user is not a member of $ADGroupName"
-                    if(Get-ADGroupMember -Identity $ADGroupName | where-Object SamAccountName -eq $user){
+                    if (Get-ADGroupMember -Identity $ADGroupName | Where-Object SamAccountName -EQ $user) {
                         throw "User: $user is already a member of AD Group: $ADGroupName"
                     }
                 }
 
                 #endregion: Validate user is not member of AD Group
 
+                <# #TODO: DELETE ME!
                 #region: Create disk using FSLogix copy-profile
+
                 $tempDiskPath = Join-PSFPath $TempFolderPath ($fslogixFilePath[0] | Split-Path -Leaf)
                 Write-PSFMessage -Level Verbose -Message "Copying Citrix profile to new VHD: $tempDiskPath"
                 New-CFXFSLogixVHD -citrixProfilePath $citrixProfilePath -Path $tempDiskPath -Username $user -FRX $FRXPath -ErrorAction Stop
 
                 #endregion: Create disk using FSLogix copy-profile
+                #>
 
-                #region: Copy redirected folders to disk using robocopy
-                Write-PSFMessage -Level Verbose -Message "Copying Redirected folders to VHD."
-                Copy-CFXRedirectedFolders -DiskPath $tempDiskPath -RedirectedFoldersPath $UserRedirectedFoldersPath -FRX $FRXPath -ErrorAction Stop
+                #region: Create disk using FSLogix create-vhd
+                $tempDiskPath = Join-PSFPath $TempFolderPath ($fslogixFilePath[0] | Split-Path -Leaf)
+                Write-PSFMessage -Level Verbose -Message " Create new VHD as: $tempDiskPath"
+                New-CFXFSLogixVHD -Path $tempDiskPath -Username $user -FRX $FRXPath -ErrorAction Stop
+                $tempDiskCreated = $true
+                #endregion: Create disk using FSLogix create-vhd
+
+                #region: Prepare disk and copy redirected folders to disk using robocopy
+                Write-PSFMessage -Level Verbose -Message "Mounting VHD using FRX.exe: $tempDiskPath"
+                $mountPoint = Mount-CFXProfile -FRX $FRXPath -DiskPath $tempDiskPath
+
+                try {
+                    $PrepareDiskError = $null
+                    Write-PSFMessage -Level Verbose -Message "Create Profile and Set permissions"
+                    Initialize-CFXFSLogixVHD -MountPointPath $mountPoint.Path -SID $sid -ErrorAction Stop
+
+                    $profilePath = Join-PSFPath $mountPoint.Path 'Profile'
+
+                    Write-PSFMessage -Level Verbose -Message "Copying Citrix UPM data"
+                    Copy-CFXData -SourcePath $citrixProfilePath -DestinationPath $profilePath -Username $user -DataLabel 'UPMProfile' -RobocopyMT $RobocopyMT -ErrorAction Stop
+
+                    Write-PSFMessage -Level Verbose -Message "Copying Redirected folders to VHD."
+                    Copy-CFXData -SourcePath $userRedirectedFoldersPath -DestinationPath $profilePath -Username $user -DataLabel 'RedirectedFolders' -RobocopyMT $RobocopyMT -ErrorAction Stop
+
+                    Write-PSFMessage -Level Verbose -Message "Creating Registry File."
+                    New-CFXFSLogixRegText -MountPointPath $mountPoint.Path -Username $user -SID $sid
+
+                }
+                catch {
+                    $PrepareDiskError = $_
+                }
+                finally {
+                    Write-PSFMessage -Level Verbose -Message "Dismounting VHD using FRX.exe"
+                    Dismount-CFXProfile -DiskPath $tempDiskPath -Cookie $mountPoint.Cookie -FRX $FRXPath
+                    if ($PrepareDiskError) { throw $PrepareDiskError }
+                }
+
                 #endregion: Copy redirected folders to disk using robocopy
 
                 #region: Copy disk to profile location(s)
@@ -140,7 +191,7 @@
                 #endregion: Copy disk to profile location(s)
 
                 #region: Add user to AD Group
-                if($ADGroupName){
+                if ($ADGroupName) {
                     Write-PSFMessage -Level Verbose -Message "Adding $user to $ADGroupName"
                     Add-ADGroupMember -Identity $ADGroupName -Members $user -ErrorAction Stop
                 }
@@ -152,12 +203,11 @@
             catch {
                 $errorMessage = $_
                 $success = $false
-                Write-PSFMessage -Message "Error" -Level Warning -ErrorRecord $_ -EnableException $true -PSCmdlet $PSCmdlet
-                Write-Error $_
+                Write-PSFMessage -Message "Error" -Level Warning -ErrorRecord $_  -PSCmdlet $PSCmdlet
             }
             finally {
                 # Delete Temp Disk
-                if(-Not $DoNoCleanup){
+                if ((-Not $DoNoCleanup) -and $tempDiskCreated) {
                     Remove-Item -Path $tempDiskPath -Force -ErrorAction Continue
                 }
 
@@ -169,8 +219,8 @@
                 Success               = $success
                 SID                   = $sid
                 CitrixProfilePath     = $citrixProfilePath
-                RedirectedFoldersPath = $UserRedirectedFoldersPath
-                VHDSizeGB             = $diskItemLength
+                RedirectedFoldersPath = $userRedirectedFoldersPath
+                VHDSizeGB             = ($diskItemLength / 1GB)
                 CopyResults           = $copyResult
                 ErrorMessage          = $errorMessage
             }
